Hello, README
This commit updates README.md with a Norwegian 101 guide to
creating pull requests, using this repository as an example.

The target audience is computer science/software engineering
students with basic experience with source control systems.

The repository was created to be used during and after a
presentation to students at Østfold University College.